Move number methods from StringUtils to a new NumberUtils trait
