+++ composer.json
@@ -19,7 +19,8 @@
     "php": ">=5.6.0 || ^7.0",
     "zendframework/zend-json": "3.0.0",
     "zf1/zend-reflection": "1.12.11",
-    "halfpastfouram/collection": "1.0.0"
+    "halfpastfouram/collection": "1.0.0",
+    "ext-json": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "5.7.*",
